+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- 3.1 - деление двух чисел
 - shared package
 - number package for shared package
 - typed_input module for shared package